+++ apps/remnant2toolkit/app/_hooks/use-pagination.ts
@@ -33,7 +33,9 @@ export function usePagination({ itemsPerPage = 16, itemsOnThisPage }: Props) {
   if (isNaN(currentPage)) currentPage = 1;
   if (currentPage < 1) currentPage = 1;
 
-  const isNextPageDisabled = itemsOnThisPage < itemsPerPage;
+  const isNextPageDisabled = useMemo(() => {
+    return itemsOnThisPage < itemsPerPage;
+  }, [itemsPerPage, itemsOnThisPage]);
 
   const firstVisibleItemNumber = useMemo(() => {
     if (itemsOnThisPage === 0) {
@@ -56,34 +58,44 @@ export function usePagination({ itemsPerPage = 16, itemsOnThisPage }: Props) {
   // If page number is less than 2, show 1 to 5
   // If the current page does not have enough items to fill the page, do not show the next pages
   const pageNumbers = useMemo(() => {
-    const numbers: number[] = [];
+    let numbers: number[] = [];
 
     if (itemsOnThisPage === 0) {
       numbers.push(1);
-      return numbers;
-    }
-
-    if (currentPage < 3) {
+    } else if (currentPage < 3) {
       for (let i = 1; i <= 5; i++) {
         numbers.push(i);
       }
-      return numbers;
-    }
-
-    if (isNextPageDisabled) {
+    } else if (isNextPageDisabled) {
       for (let i = currentPage - 4; i <= currentPage; i++) {
         if (i > 0) {
           numbers.push(i);
         }
       }
-      return numbers;
-    }
-
-    for (let i = currentPage - 2; i <= currentPage + 2; i++) {
-      if (i > 0) {
-        numbers.push(i);
+    } else {
+      for (let i = currentPage - 2; i <= currentPage + 2; i++) {
+        if (i > 0) {
+          numbers.push(i);
+        }
       }
     }
+
+    // If next page is disabled, do not show the next pages
+    if (isNextPageDisabled) {
+      numbers = numbers.filter((number) => number <= currentPage);
+      // If the total page count is less than 5, and the first page is greater than 1, add
+      // pages to the start of the list until total count is 5
+      if (numbers.length < 5) {
+        const firstPage = numbers[0] as number;
+        for (let i = firstPage - 1; i > 0; i--) {
+          numbers.unshift(i);
+          if (numbers.length === 5) {
+            break;
+          }
+        }
+      }
+    }
+
     return numbers;
   }, [currentPage, isNextPageDisabled, itemsOnThisPage]);