+++ packages/ui/src/common/icons/new-window/index.tsx
@@ -1,6 +1,6 @@
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faWindowMaximize } from '@fortawesome/free-solid-svg-icons/faWindowMaximize';
+import { faUpRightFromSquare } from '@fortawesome/free-solid-svg-icons/faUpRightFromSquare';
 
 export function NewWindowIcon({ className }: { className?: string }) {
-  return <FontAwesomeIcon className={className} icon={faWindowMaximize} />;
+  return <FontAwesomeIcon className={className} icon={faUpRightFromSquare} />;
 }