+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## 2024-10-26
+
+### Fixed
+
+- Fixed some inconsistencies in pagination in build lists.
+
 ## 2024-10-25
 
 ### Added